# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/collections": {
            "get": {
                "description": "Collection列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Collection列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "父级ID",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "获取深度",
                        "name": "depth",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Collection",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否包含total",
                        "name": "include_total",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pre_page",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "展开数据",
                        "name": "expand",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ListCollectionResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Create Collection",
                "parameters": [
                    {
                        "description": "Collection data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateCollectionParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            }
        },
        "/collections/{id}": {
            "get": {
                "description": "Get Collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Get Collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "expand attribute list",
                        "name": "expand",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            },
            "post": {
                "description": "Update Collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Update Collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateCollectionParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Collection"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Delete Collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CollectionID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HttpError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Authorize": {
            "type": "object",
            "properties": {
                "attribute": {
                    "description": "鉴权属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/value.AuthorizeAttribute"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "鉴权备注",
                    "type": "string"
                },
                "type": {
                    "description": "鉴权类型",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.Collection": {
            "type": "object",
            "properties": {
                "authorize": {
                    "description": "鉴权信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.Authorize"
                        }
                    ]
                },
                "authorize_id": {
                    "description": "鉴权信息ID",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endpoint": {
                    "description": "后端服务\n集合中没有设置后端服务的路由默认继承集合的后端服务信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.Endpoint"
                        }
                    ]
                },
                "endpoint_id": {
                    "description": "后端服务",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "string"
                },
                "routes": {
                    "description": "路由信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Route"
                    }
                },
                "server_names": {
                    "description": "服务域名\n外部服务访问路由的入口",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.Endpoint": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "服务描述",
                    "type": "string"
                },
                "endpoint": {
                    "description": "远程服务",
                    "allOf": [
                        {
                            "$ref": "#/definitions/value.ForwardEndpoint"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "后端名",
                    "type": "string"
                },
                "type": {
                    "description": "服务类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.EndpointType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.Route": {
            "type": "object",
            "properties": {
                "authorize": {
                    "description": "鉴权信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.Authorize"
                        }
                    ]
                },
                "authorize_id": {
                    "description": "鉴权信息ID",
                    "type": "string"
                },
                "collection_id": {
                    "description": "分组ID",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endpoint": {
                    "description": "路由自定义的后端路由",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.Endpoint"
                        }
                    ]
                },
                "endpoint_id": {
                    "description": "后端服务",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "match_options": {
                    "description": "路由的特殊匹配规则",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/value.MatchOption"
                    }
                },
                "method": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "modify_options": {
                    "description": "数据重写规则",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/value.ModifyOption"
                    }
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "description": "路径",
                    "type": "string"
                },
                "path_rewrite": {
                    "description": "路由重写规则",
                    "allOf": [
                        {
                            "$ref": "#/definitions/value.PathRewrite"
                        }
                    ]
                },
                "path_type": {
                    "description": "路径类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.RoutePathType"
                        }
                    ]
                },
                "status": {
                    "description": "状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.RouteStatus"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "enum.EndpointType": {
            "type": "string",
            "enum": [
                "static"
            ],
            "x-enum-varnames": [
                "EndpointTypeStatic"
            ]
        },
        "enum.RoutePathType": {
            "type": "string",
            "enum": [
                "exact",
                "prefix"
            ],
            "x-enum-varnames": [
                "RoutePathTypeExact",
                "RoutePathTypePrefix"
            ]
        },
        "enum.RouteStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive"
            ],
            "x-enum-varnames": [
                "RouteStatusActive",
                "RouteStatusInactive"
            ]
        },
        "httputil.HttpError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httputil.HttpErrorDetail"
                }
            }
        },
        "httputil.HttpErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "service.CreateCollectionParam": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "authorize_id": {
                    "description": "鉴权配置",
                    "type": "string"
                },
                "description": {
                    "description": "分组描述",
                    "type": "string"
                },
                "endpoint_id": {
                    "description": "绑定的后端服务",
                    "type": "string"
                },
                "name": {
                    "description": "分组名",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级节点",
                    "type": "string"
                },
                "server_names": {
                    "description": "绑定的域名",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "service.ListCollectionResult": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Collection"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.UpdateCollectionParam": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "authorize_id": {
                    "description": "鉴权配置",
                    "type": "string"
                },
                "description": {
                    "description": "分组描述",
                    "type": "string"
                },
                "endpoint_id": {
                    "description": "绑定的后端服务",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "分组名",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级节点",
                    "type": "string"
                },
                "server_names": {
                    "description": "绑定的域名",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "value.AuthorizeAttribute": {
            "type": "object",
            "properties": {
                "binary": {
                    "$ref": "#/definitions/value.AuthorizeAttributeBinary"
                }
            }
        },
        "value.AuthorizeAttributeBinary": {
            "type": "object",
            "required": [
                "sources"
            ],
            "properties": {
                "header": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/value.AuthorizeSource"
                    }
                }
            }
        },
        "value.AuthorizeSource": {
            "type": "object",
            "required": [
                "name",
                "source"
            ],
            "properties": {
                "name": {
                    "description": "匹配值",
                    "type": "string"
                },
                "source": {
                    "description": "匹配源",
                    "type": "string"
                }
            }
        },
        "value.ForwardEndpoint": {
            "type": "object",
            "properties": {
                "static": {
                    "$ref": "#/definitions/value.ForwardEndpointStatic"
                }
            }
        },
        "value.ForwardEndpointStatic": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/value.ForwardEndpointTarget"
                    }
                },
                "timeout": {
                    "type": "integer"
                }
            }
        },
        "value.ForwardEndpointTarget": {
            "type": "object",
            "required": [
                "address",
                "network",
                "weight"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
        "value.MatchOption": {
            "type": "object",
            "required": [
                "name",
                "source",
                "type",
                "value"
            ],
            "properties": {
                "name": {
                    "description": "匹配值",
                    "type": "string"
                },
                "source": {
                    "description": "匹配源",
                    "type": "string"
                },
                "type": {
                    "description": "匹配方式",
                    "type": "string"
                },
                "value": {
                    "description": "匹配内容",
                    "type": "string"
                }
            }
        },
        "value.ModifyOption": {
            "type": "object",
            "required": [
                "name",
                "source",
                "type"
            ],
            "properties": {
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "source": {
                    "description": "编辑源",
                    "type": "string"
                },
                "type": {
                    "description": "编辑类型",
                    "type": "string"
                },
                "value": {
                    "description": "值",
                    "type": "string"
                }
            }
        },
        "value.PathRewrite": {
            "type": "object",
            "properties": {
                "regex": {
                    "type": "string"
                },
                "replace": {
                    "type": "string"
                }
            }
        }
    }
}